requirements.txt
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-language==1.3.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.12
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
